requirements.txt
asgiref==3.5.0
certifi==2021.10.8
dj-database-url==0.5.0
Django==4.0.2
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-js-asset==2.0.0
gunicorn==20.1.0
Pillow==9.0.1
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.0.0
